third_party/WebKit/LayoutTests/fast/block/float/relayout-nested-float-after-line.html

<!DOCTYPE html>
<p>The word "BONZER" should be adjacent to the hotpink square.</p>
<div style="position:relative; overflow:hidden;">
    &nbsp;
    <div style="float:left; width:100%; height:1px;"></div>
    <div style="float:left; width:100%;">
        <div id="innerFloat" style="float:left; width:666px; height:20px; background:hotpink;"></div>
        <span id="targetSpan">BONZER<br></span>
    </div>
</div>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
test(() => {
    document.body.offsetTop;
    innerFloat.style.width = "20px";
    assert_equals(targetSpan.offsetLeft, 20);
}, "Resize inner float next to line");
</script>
